requirements.txt
python-telegram-bot==13.5
requests
Pillow==8.3.1
pymongo==3.10.0
